Add Firebase configuration files and update .gitignore; refactor imports to use ES6 module syntax across multiple files
