+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "news-api-f1a79"
+  }
+}
