+++ firebase.json
@@ -1,0 +1,5 @@
+{
+  "emulators": {
+    "singleProjectMode": true
+  }
+}
