+++ rss functions/functions/routes/social.js
@@ -3,7 +3,7 @@ import { Router } from 'express';
 // eslint-disable-next-line new-cap
 const router = Router();
 
-import { processAndStoreSocialPosts } from '../services/socialService';
+import { processAndStoreSocialPosts } from '../services/socialService.js';
 
 // GET /social - Process and store social media posts
 router.get('/social', async (req, res) => {
